[jitendra] - Version-3 has been implemented
What changes have you made?

added main class called Label, and RecordsFormatter for format the records,
People for filter and print.

Why do we need this change?

We should be able to filter the user according to given criteria and after
filter we should be able to print it.

What can this change impact?
 Changes to be committed:
	new file:   TODO
	modified:   person.txt
	new file:   src/Address.java
	new file:   src/Age.java
	new file:   src/City.java
	new file:   src/CliOptions.java
	new file:   src/Country.java
	modified:   src/Gender.java
	new file:   src/Label.java
	new file:   src/LabelPrinter.java
	deleted:    src/Main.java
	modified:   src/Name.java
	new file:   src/People.java
	modified:   src/Person.java
	deleted:    src/Place.java
	new file:   src/RecordsFormatter.java
	new file:   src/State.java
	new file:   test/CliOptionsTest.java
	new file:   test/NameTest.java
	new file:   test/PeopleTest.java
	modified:   test/PersonTest.java
	new file:   test/RecordsFormatterTest.java